Reconfigure ThreatStack agent automatically if rulesets for the node change
According to TS support, there is no easy way to obtain a list of configured rulesets
for a node. Add a file to store the list and reconfigure the agent when the list changes.
As part of reconfiguration ThreatStack agent needs to be stopped. The reconfiguration
will cause a stale TS node with the same name to show in the UI for a while.